requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asyncio-extras==1.3.2
asyncpraw==7.6.1
asyncprawcore==2.3.0
attrs==22.1.0
booru==1.0.15
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord==2.1.0
discord.py==2.1.0
exceptiongroup==1.0.4
Flask==2.2.2
frozenlist==1.3.3
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
pluggy==1.0.0
praw==7.6.1
prawcore==2.3.0
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.2
python-dotenv==0.21.0
requests==2.28.1
thor==0.9.6
tomli==2.0.1
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.13
waitress==2.1.2
websocket-client==1.4.2
Werkzeug==2.2.2
xmltodict==0.13.0
yarl==1.8.1
